requirements.txt
python-dotenv
mcp-python>=0.3.0
groq>=0.4.0
aiohttp>=3.9.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
